==> 5_DownloadImages.ipynb <==
# Setup

import pandas as pd
import numpy as np
import requests
from PIL import Image
import shutil
import urllib

# Import Dataset

df = pd.read_csv("../NespressoTrainingApp/PreparedCoffeeData.csv");
imgCols = ['ID','Name','Type','Capsule Image Link','Capsule & Sleeve Image Link'];
df[imgCols].head()

# Download Images

* https://blog.finxter.com/5-easy-ways-to-download-an-image-from-a-url-in-python/
* https://www.w3schools.com/python/python_try_except.asp

def download_image(dataframe, index, folder):
    ID = dataframe.loc[index, "ID"];
    Name = dataframe.loc[index, "Name"];
    path = f"../NespressoTrainingApp/assets/{folder}/{ID} {Name}";
    Type = dataframe.loc[index, "Type"];
    success = True;
    
    # Capsule Image ------------
    url_capsule = dataframe.loc[index, "Capsule Image Link"];
    path_capsule = path + " Capsule.png";
    
    # Capsule & Sleeve ----------
    url_capsuleAndSleeve = dataframe.loc[index, "Capsule & Sleeve Image Link"];
    path_capsuleAndSleeve = path + " CapsuleAndSleeve.png";

    try:
        img_capsule = Image.open(requests.get(url_capsule, stream = True).raw);
        img_capsuleAndSleeve = Image.open(requests.get(url_capsuleAndSleeve, stream = True).raw);
    except:
        success = False;
    
    if success == False:
        print(Type + " " + Name + " coffee images were unable to be downloaded. The images need to be downloaded manually.");
    else:
        img_capsule.save(path_capsule);
        img_capsuleAndSleeve.save(path_capsuleAndSleeve);
        dataframe.loc[index, "Capsule Image Link"] = path_capsule[24:];
        dataframe.loc[index, "Capsule & Sleeve Image Link"] = path_capsuleAndSleeve[24:];
        

for Type in df['Type'].unique().tolist():
    for i in df[df['Type'] == Type].index:
        download_image(df, i, Type.lower());

napoli_id = df[df['Name'] == 'Inspirazione Napoli'].index[0].astype(int);
print(napoli_id);
df.loc[napoli_id, "Capsule Image Link"] = "assets/original/OL1 Napoli Capsule.png";
df.loc[napoli_id, "Capsule & Sleeve Image Link"] = "assets/original/OL1 Napoli CapsuleAndSleeve.png";

chiaro_id = df[df['Name'] == "Chiaro"].index[0].astype(int);
print(chiaro_id);
df.loc[chiaro_id, "Capsule Image Link"] = "assets/original/OL32 Chiaro Capsule.png";
df.loc[chiaro_id, "Capsule & Sleeve Image Link"] = "assets/original/OL32 Chiaro CapsuleAndSleeve.png";

df[imgCols]

df.to_csv("../NespressoTrainingApp/PreparedCoffeeData_Local.csv", index=False)
